+++ UltitemsCyan/Items/Tier3/ViralEssence.cs
@@ -12,7 +12,8 @@ namespace UltitemsCyan.Items.Tier3
     public class ViralEssence : ItemBase
     {
         public static ItemDef item;
-        private const float speedPerStackStatus = 20f;
+        private const float speedPerStackStatus = 16f;
+        private const int debuffMultplier = 2;
         private const int maxStatusCountperItem = 6;
 
         public override void Init(ConfigFile configs)
@@ -26,7 +27,7 @@ namespace UltitemsCyan.Items.Tier3
                 "VIRALESSENCE",
                 itemName,
                 "Increase speed per unique status effect.",
-                "Increases <style=cIsUtility>movement speed</style> by <style=cIsUtility>20%</style> <style=cStack>(+20% per stack)</style> per <style=cIsDamage>unique status</style> you have. Up to a maximum of <style=cIsUtility>6</style> <style=cStack>(+6 per stack)</style>.",
+                "Increases <style=cIsUtility>movement speed</style> by <style=cIsUtility>16%</style> <style=cStack>(+16% per stack)</style> per <style=cIsDamage>unique status</style> you have. <style=cIsUtility>Twice</style> as much for negative status. Up to a maximum of <style=cIsUtility>6</style> <style=cStack>(+6 per stack)</style>.",
                 "Illness",
                 ItemTier.Tier3,
                 UltAssets.ViralSmogSprite,
@@ -58,29 +59,45 @@ namespace UltitemsCyan.Items.Tier3
                     BuffIndex[] activeBuffs = sender.activeBuffsList;
                     //Log.Debug("Active Count Length: " + activeBuffLength);
 
-                    int nonCooldownBuffs = activeBuffLength;
+                    //int nonCooldownBuffs = activeBuffLength;
+
+                    int debuffCount = 0;
+                    int buffCount = 0;
+                    
+
+                    //Log.Debug("Counting Status ------------------------------ ");
                     for (int i = 0; i < activeBuffLength; i++)
                     {
                         BuffDef buffDef = BuffCatalog.GetBuffDef(activeBuffs[i]);
-                        if (buffDef.isCooldown)
+                        if (buffDef.isDebuff || buffDef.isDOT)
                         {
-                            nonCooldownBuffs--;
-                            //Log.Debug(" ~ Cooldown: " + buffDef.name + "\tnew nonCool Count: " + nonCooldownBuffs);
+                            //Log.Debug("De Buff:  " + buffDef.name);
+                            debuffCount++;
+                            if (debuffCount >= maxStatusCountperItem * grabCount)
+                            {
+                                // Counted max multiplier
+                                break;
+                            }
                         }
-                        else
+                        else if (!buffDef.isCooldown && !buffDef.isHidden)
                         {
-                            //Log.Debug("Not Cool: " + buffDef.name);
+                            //Log.Debug("Pos Buff: " + buffDef.name);
+                            buffCount++;
                         }
                     }
 
-                    nonCooldownBuffs = Math.Min(nonCooldownBuffs, maxStatusCountperItem * grabCount);
+                    int statusTotal = Math.Min(debuffCount, maxStatusCountperItem * grabCount) * debuffMultplier +
+                                      Math.Min(buffCount, maxStatusCountperItem * grabCount - debuffCount);
+
+                    //Log.Debug("Total !!! " + Math.Min(debuffCount, maxStatusCountperItem * grabCount) + " * 2 + " +
+                    //            Math.Min(buffCount, maxStatusCountperItem * grabCount - debuffCount) + " = " + statusTotal);
 
                     //Log.Debug("Viral Smog\nCount: " + nonCooldownBuffs + "\n"
                     //    + "Speed from Virus: " + speedPerStackStatus / 100f * nonCooldownBuffs * grabCount);
                     // Gives 20% speed per status per item
                     if (activeBuffLength > 0)
                     {
-                        args.moveSpeedMultAdd += speedPerStackStatus / 100f * nonCooldownBuffs * grabCount;
+                        args.moveSpeedMultAdd += speedPerStackStatus / 100f * statusTotal * grabCount;
                     }
                     else
                     {