+++ UltitemsCyan/Items/Void/DownloadedRAM.cs
@@ -1,9 +1,10 @@
-﻿using RoR2;
+﻿using BepInEx.Configuration;
+using RoR2;
 using System;
 using UltitemsCyan.Buffs;
 using UltitemsCyan.Items.Tier2;
 using UnityEngine;
-using BepInEx.Configuration;
+using UnityEngine.AddressableAssets;
 using UnityEngine.Networking;
 
 namespace UltitemsCyan.Items.Void
@@ -16,11 +17,18 @@ namespace UltitemsCyan.Items.Void
     // TODO: check if Item classes needs to be public
     public class DownloadedRAM : ItemBase
     {
+
+        //"RoR2/Base/Imp/ImpDeathEffect.prefab" (like a dark puddle)
+        public readonly GameObject ImpEffect = Addressables.LoadAssetAsync<GameObject>("RoR2/DLC1/EliteVoid/VoidInfestorLeapEffect.prefab").WaitForCompletion();
+
         public static ItemDef item;
         public static ItemDef transformItem;
 
-        public const float downloadedBuffMultiplier = 8f;
+        public const float downloadedBuffMultiplier = 20f;
         public const int downloadsPerItem = 4;
+
+        public const float plugeSpeed = 8f;
+        public const float plugeSpeedDivisor = 2f;
 
         private const float downloadChance = 12f;
 
@@ -33,11 +41,12 @@ namespace UltitemsCyan.Items.Void
             {
                 return;
             }
+            //<style=cIsUtility>plunge</style> enemies and 
             item = CreateItemDef(
                 "DOWNLOADEDRAM",
                 itemName,
                 "Chance on hit to temporarily increase damage. Dealing damage refreshes the timer. <style=cIsVoid>Corrupts all Overclocked GPUs</style>.",
-                "<style=cIsDamage>12%</style> chance on hit to increase damage by <style=cIsDamage>8%</style>. Maxinum cap of <style=cIsDamage>4</style> <style=cStack>(+4 per stack)</style>. Lose stacks 4 seconds after not inflicting damage. <style=cIsVoid>Corrupts all Overclocked GPUs</style>.",
+                "<style=cIsDamage>12%</style> chance on hit to increase your damage by <style=cIsDamage>8%</style>, up to <style=cIsDamage>4</style> <style=cStack>(+4 per stack)</style>, for 4s. Dealing damage refreshes the timer. <style=cIsVoid>Corrupts all Overclocked GPUs</style>.",
                 "Wow I can't belive it worked! I thought for sure it was a scam!",
                 ItemTier.VoidTier2,
                 UltAssets.DownloadedRAMSprite,
@@ -62,30 +71,6 @@ namespace UltitemsCyan.Items.Void
                 _ = self.AddItemBehavior<DownloadedVoidBehavior>(self.inventory.GetItemCountEffective(item));
             }
         }
-
-        /*/
-        private void CharacterBody_OnOutOfDangerChanged(On.RoR2.CharacterBody.orig_OnOutOfDangerChanged orig, CharacterBody self)
-        {
-            Log.Warning(" ! Combat Changed ! ");
-            try
-            {
-                if (self && self.outOfCombat)
-                {
-                    Log.Debug(self.name + "  ...Leaving Combat");
-                    self.SetBuffCount(DownloadedBuff.buff.buffIndex, 0);
-                }
-                else
-                {
-                    Log.Debug(self.name + " is Entering Combat ! ! !");
-                }
-            }
-            catch (NullReferenceException)
-            {
-                Log.Debug("Who Downloaded?");
-                Log.Debug("Name: " + self.name);
-            }
-        }
-        //*/
 
         private void GlobalEventManager_OnHitEnemy(On.RoR2.GlobalEventManager.orig_OnHitEnemy orig, GlobalEventManager self, DamageInfo damageInfo, GameObject victim)
         {
@@ -112,6 +97,34 @@ namespace UltitemsCyan.Items.Void
                         if (Util.CheckRoll(downloadChance, inflictor.master.luck))
                         {
                             //Log.Debug("downloading");
+
+                            // Launch victim downwards
+                            CharacterBody victimBody = victim.GetComponent<CharacterBody>();
+                            //victimBody.characterMotor.ModifyGravity()
+                            if (victimBody != null && victimBody.characterMotor)
+                            {
+                                //ref Vector3 refVelocity = ref victimBody.characterMotor.velocity;
+                                //Log.Debug("Vertical Velocity: initial " + refVelocity.y);
+                                //if (refVelocity.y >= 0f)
+                                //{
+                                //    refVelocity.y /= plugeSpeedDivisor;
+                                //}
+                                //refVelocity.x /= plugeSpeedDivisor;
+                                //refVelocity.z /= plugeSpeedDivisor;
+
+                                //refVelocity.y -= plugeSpeed;
+
+                                //Log.Debug("                       new " + refVelocity.y);
+                            }
+
+                            EffectManager.SpawnEffect(ImpEffect, new EffectData
+                            {
+                                origin = victimBody.corePosition,
+                                //color = new Color(0.2392f, 0.8196f, 0.917647f) // Cyan Lunar color
+                            }, true);
+
+
+
                             // If you have fewer than the max number of downloads, then grant buff
                             if (inflictor.GetBuffCount(DownloadedBuff.buff) < grabCount * downloadsPerItem)
                             {